메인 컴포넌트 분리 완료, 바텀 네비게이션 만드는 중
