+++ lib/holder/main/_auto_ad_slider.dart
