+++ lib/holder/main/_creator_scroll.dart
@@ -1,0 +1,31 @@
+import 'package:flutter/cupertino.dart';
+import 'package:flutter/material.dart';
+
+import '../../models/creator.dart';
+import '_creator_circle.dart';
+
+class CreatorScroll extends StatelessWidget {
+  const CreatorScroll({
+    super.key,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return SliverToBoxAdapter(
+      child: SizedBox(
+        height: 70,
+        child: ListView.builder(
+          scrollDirection: Axis.horizontal,
+          itemCount: creatorList.length,
+          itemBuilder: (context, index) {
+            /* 콜백 함수, 각 index에 자식 위젯 빌드 */ {
+              return CreatorCircle(
+                creator: creatorList[index],
+              );
+            }
+          },
+        ),
+      ),
+    );
+  }
+}
