+++ lib/main.dart
@@ -1,7 +1,7 @@
 import 'package:final_project_team02/theme.dart';
 import 'package:flutter/material.dart';
 
-import 'holder/main_holder.dart';
+import 'holder/main/main_holder.dart';
 
 void main() {
   runApp(const MyApp());
